Move FocusCapture to own file
