+++ packages/block-editor/src/components/writing-flow/index.js
@@ -6,7 +6,7 @@ import { overEvery, find, findLast, reverse, first, last } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { forwardRef, useRef } from '@wordpress/element';
+import { useRef } from '@wordpress/element';
 import {
 	computeCaretRect,
 	focus,
@@ -29,6 +29,7 @@ import {
 	hasInnerBlocksContext,
 	getBlockFocusableWrapper,
 } from '../../utils/dom';
+import FocusCapture from './focus-capture';
 
 /**
  * Browser constants
@@ -71,83 +72,6 @@ export function isNavigationCandidate( element, keyCode, hasModifier ) {
 	const { tagName } = element;
 	return tagName !== 'INPUT' && tagName !== 'TEXTAREA';
 }
-
-/**
- * Renders focus capturing areas to redirect focus to the selected block if not
- * in Navigation mode.
- *
- * @param {string}  selectedClientId Client ID of the selected block.
- * @param {boolean} isReverse        Set to true if the component is rendered
- *                                   after the block list, false if rendered
- *                                   before.
- * @param {Object}  containerRef     Reference containing the element reference
- *                                   of the block list container.
- * @param {boolean} noCapture        Reference containing the flag for enabling
- *                                   or disabling capturing.
- *
- * @return {WPElement} The focus capture element.
- */
-const FocusCapture = forwardRef( ( {
-	selectedClientId,
-	isReverse,
-	containerRef,
-	noCapture,
-}, ref ) => {
-	const isNavigationMode = useSelect( ( select ) =>
-		select( 'core/block-editor' ).isNavigationMode()
-	);
-	const { setNavigationMode } = useDispatch( 'core/block-editor' );
-
-	function onFocus() {
-		// Do not capture incoming focus if set by us in WritingFlow.
-		if ( noCapture.current ) {
-			delete noCapture.current;
-			return;
-		}
-
-		// When focus coming in from out of the block list, and no block is
-		// selected, enable Navigation mode and select the first or last block
-		// depending on the direction.
-		if ( ! selectedClientId ) {
-			setNavigationMode( true );
-
-			const tabbables = focus.tabbable.find( containerRef.current );
-
-			if ( tabbables.length ) {
-				if ( isReverse ) {
-					last( tabbables ).focus();
-				} else {
-					first( tabbables ).focus();
-				}
-			}
-
-			return;
-		}
-
-		// If there is a selected block, move focus to the first or last
-		// tabbable element depending on the direction.
-		const wrapper = getBlockFocusableWrapper( selectedClientId );
-
-		if ( isReverse ) {
-			const tabbables = focus.tabbable.find( wrapper );
-			last( tabbables ).focus();
-		} else {
-			wrapper.focus();
-		}
-	}
-
-	return (
-		<div
-			ref={ ref }
-			// Don't allow tabbing to this element in Navigation mode.
-			tabIndex={ ! isNavigationMode ? '0' : undefined }
-			onFocus={ onFocus }
-			// Needs to be positioned within the viewport, so focus to this
-			// element does not scroll the page.
-			style={ { position: 'fixed' } }
-		/>
-	);
-} );
 
 /**
  * Returns the optimal tab target from the given focused element in the
